Module Raylib.Matrix

type t'

type t = t' ctyp

val create : float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> float -> t

create m0 m1 m2 m3 m4 m5 m6 m7 m8 m9 m10 m11 m12 m13 m14 m15

val m0 : t -> float

val m1 : t -> float

val m2 : t -> float

val m3 : t -> float

val m4 : t -> float

val m5 : t -> float

val m6 : t -> float

val m7 : t -> float

val m8 : t -> float

val m9 : t -> float

val m10 : t -> float

val m11 : t -> float

val m12 : t -> float

val m13 : t -> float

val m14 : t -> float

val m15 : t -> float

val set_m0 : t -> float -> unit

val set_m1 : t -> float -> unit

val set_m2 : t -> float -> unit

val set_m3 : t -> float -> unit

val set_m4 : t -> float -> unit

val set_m5 : t -> float -> unit

val set_m6 : t -> float -> unit

val set_m7 : t -> float -> unit

val set_m8 : t -> float -> unit

val set_m9 : t -> float -> unit

val set_m10 : t -> float -> unit

val set_m11 : t -> float -> unit

val set_m12 : t -> float -> unit

val set_m13 : t -> float -> unit

val set_m14 : t -> float -> unit

val set_m15 : t -> float -> unit

val determinant : t -> float

val trace : t -> float

val transpose : t -> t

val invert : t -> t

val normalize : t -> t

val identity : unit -> t

val add : t -> t -> t

val subtract : t -> t -> t

val translate : float -> float -> float -> t

val rotate : Vector3.t -> float -> t

val rotate_xyz : Vector3.t -> t

val rotate_zyx : Vector3.t -> t

val rotate_x : float -> t

val rotate_y : float -> t

val rotate_z : float -> t

val scale : float -> float -> float -> t

val multiply : t -> t -> t

val frustum : float -> float -> float -> float -> float -> float -> t

frustum left right bottom top near far returns perspective projection matrix

val perspective : float -> float -> float -> float -> t

perspective fovy aspect near far returns perspective projection matrix

val ortho : float -> float -> float -> float -> float -> float -> t

ortho left right bottom top near far returns orthographic projection matrix

val look_at : Vector3.t -> Vector3.t -> Vector3.t -> t

look at eye target up returns camera look-at matrix (view matrix)